+++ libcdpfgl/files.h
@@ -267,7 +267,7 @@ extern gchar *normalize_directory(gchar *path);
  * @returns the maximum number of hashs that may be asked for into a
  *          single GET HTTP request.
  */
-extern gint64 calculate_max_number_of_hashs(gint64 size);
+extern gint calculate_max_number_of_hashs(guint64 size);
 
 
 #endif /* #ifndef _FILES_H_ */
